升级JDK版本到1.8, 支持字段查询和更新的排除

+++ myframe-generator/src/main/java/com/myframe/generator/util/DbUtils.java
@@ -43,8 +43,8 @@ public class DbUtils {
     static {
         typeMap.put("BIGINT", "Long");
         typeMap.put("INTEGER", "Integer");
-        typeMap.put("TINYINT", "Short");
-        typeMap.put("BIT", "Short");
+        typeMap.put("TINYINT", "Integer");
+        typeMap.put("BIT", "Integer");
         typeMap.put("SMALLINT", "Integer");
         typeMap.put("FLOAT", "Float");
         typeMap.put("DOUBLE", "Double");
